requirements.txt
Flask==3.0.0
nltk==3.8.1
scikit-learn==1.4.0
pymongo==4.6.1
python-docx==1.1.0
PyPDF2==3.0.1
transformers==4.41.2
torch==2.3.0
